Refactor backend configuration: remove WebMvcConfigurer, add SpringConfiguration for CORS and resource handling; update .gitignore and Vite build output directory
