+++ .gitignore
@@ -1,1 +1,2 @@
 pnwcsclub-backend/src/main/resources/application.properties
+pnwcsclub-backend/src/main/resources/static
