+++ pnwcsclub-backend/src/main/java/website/pnwcsclub_backend/configuration/SpringConfiguration.java
@@ -1,0 +1,53 @@
+package website.pnwcsclub_backend.configuration;
+import java.util.List;
+import static java.util.Objects.nonNull;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.Resource;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.resource.PathResourceResolver;
+import org.springframework.web.servlet.resource.ResourceResolverChain;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+@Configuration
+public class SpringConfiguration implements WebMvcConfigurer {
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/api/**")
+                .allowedOrigins("http://localhost:5173") //no work, supposed to allow requests from this origin
+                .allowedMethods("GET", "POST", "PUT", "DELETE")
+                .allowedHeaders("*");
+    };          
+
+ 
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        this.serveDirectory(registry, "/", "classpath:/static/");
+    }
+ 
+    private void serveDirectory(ResourceHandlerRegistry registry, String endpoint, String location) {
+         String[] endpointPatterns = endpoint.endsWith("/")
+            ? new String[]{endpoint.substring(0, endpoint.length() - 1), endpoint, endpoint + "**"}
+            : new String[]{endpoint, endpoint + "/", endpoint + "/**"};
+    registry
+            // 2
+            .addResourceHandler(endpointPatterns)
+            .addResourceLocations(location.endsWith("/") ? location : location + "/")
+            .resourceChain(false)
+            // 3
+            .addResolver(new PathResourceResolver() {
+                @Override
+                public Resource resolveResource(HttpServletRequest request, String requestPath, List<? extends Resource> locations, ResourceResolverChain chain) {
+                    Resource resource = super.resolveResource(request, requestPath, locations, chain);
+                    if (nonNull(resource)) {
+                        return resource;
+                    }
+                    return super.resolveResource(request, "/index.html", locations, chain);
+                }
+            });
+    }
+ 
+}
