--- pnwcsclub-backend/src/main/java/website/pnwcsclub_backend/configuration/WebMvcConfigurer.java
@@ -1,14 +1,0 @@
-package website.pnwcsclub_backend.configuration;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-
-@Configuration
-public class WebMvcConfigurer {
-    //@Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/api/**")
-                .allowedOrigins("http://localhost:5173") //no work, supposed to allow requests from this origin
-                .allowedMethods("GET", "POST", "PUT", "DELETE")
-                .allowedHeaders("*");
-    };          
-}
