+++ pnwcsclub-front/vite.config.js
@@ -4,4 +4,7 @@ import react from '@vitejs/plugin-react'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  build: {
+    outDir: '../pnwcsclub-backend/src/main/resources/static'
+  }
 })
